Move copy action out of the controller and handle the no post found error
startEditingPostCopy is moved along the other Flux actions, waiting to be replaced by its Redux version.

If the post ID does not exist, the attempt to copy it quietly fails into a new empty draft.

See:
https://github.com/Automattic/wp-calypso/pull/7451#discussion_r74938635
https://github.com/Automattic/wp-calypso/pull/7451#discussion_r74940270
