+++ client/lib/posts/actions.js
@@ -2,11 +2,8 @@
  * External dependencies
  */
 var debug = require( 'debug' )( 'calypso:posts' ),
-	defer = require( 'lodash/defer' ),
-	store = require( 'store' ),
-	clone = require( 'lodash/clone' ),
-	fromPairs = require( 'lodash/fromPairs' ),
-	assign = require( 'lodash/assign' );
+	store = require( 'store' );
+import { assign, clone, defer, fromPairs, map, pick } from 'lodash';
 
 /**
  * Internal dependencies
@@ -147,6 +144,42 @@ PostActions = {
 				error: error,
 				post: data
 			} );
+		} );
+	},
+
+	/**
+	 * Copy an existing post into a new draft and start keeping track of edits to it
+	 *
+	 * @param {Number} siteId Site ID to copy post from
+	 * @param {Number} postId Post ID to copy
+	 */
+	startEditingPostCopy( siteId, postId ) {
+		if ( ! siteId ) {
+			return;
+		}
+
+		wpcom.site( siteId ).post( postId ).get().then( post => {
+			this.startEditingNew( siteId, {
+				type: 'post',
+				content: post.content,
+			} );
+
+			const postAttributes = pick(
+				post,
+				'canonical_image',
+				'excerpt',
+				'featured_image',
+				'format',
+				'metadata',
+				'post_thumbnail',
+				'title'
+			);
+			postAttributes.categories = map( post.categories, 'ID' );
+			postAttributes.tags = map( post.tags, 'name' );
+
+			this.edit( postAttributes );
+		} ).catch( () => {
+			this.startEditingNew( siteId, { type: 'post' } );
 		} );
 	},
